refactor

+++ solvexity/trader/model/kline.py
@@ -1,19 +1,20 @@
 from pydantic import BaseModel
+from decimal import Decimal
 
 class KLine(BaseModel):
     interval: str
     open_time: int
     close_time: int
     event_time: int
-    open: float
-    high: float
-    low: float
-    close: float
+    open: Decimal
+    high: Decimal
+    low: Decimal
+    close: Decimal
     number_of_trades: int
-    base_asset_volume: float
-    quote_asset_volume: float
-    taker_buy_base_asset_volume: float
-    taker_buy_quote_asset_volume: float
+    base_asset_volume: Decimal
+    quote_asset_volume: Decimal
+    taker_buy_base_asset_volume: Decimal
+    taker_buy_quote_asset_volume: Decimal
     is_closed: bool
 
     @classmethod
@@ -23,15 +24,15 @@ class KLine(BaseModel):
             open_time=data["t"],
             close_time=data["T"],
             event_time=event_time,
-            open=float(data["o"]),
-            high=float(data["h"]),
-            low=float(data["l"]),
-            close=float(data["c"]),
+            open=Decimal(data["o"]),
+            high=Decimal(data["h"]),
+            low=Decimal(data["l"]),
+            close=Decimal(data["c"]),
             number_of_trades=data["n"],
-            base_asset_volume=float(data["v"]),
-            quote_asset_volume=float(data["q"]),
-            taker_buy_base_asset_volume=float(data["V"]),
-            taker_buy_quote_asset_volume=float(data["Q"]),
+            base_asset_volume=Decimal(data["v"]),
+            quote_asset_volume=Decimal(data["q"]),
+            taker_buy_base_asset_volume=Decimal(data["V"]),
+            taker_buy_quote_asset_volume=Decimal(data["Q"]),
             is_closed=data["x"]
         )
     
@@ -42,14 +43,14 @@ class KLine(BaseModel):
             open_time=int(data[0]),
             close_time=int(data[6]),
             event_time=int(data[6]),
-            open=float(data[1]),
-            high=float(data[2]),
-            low=float(data[3]),
-            close=float(data[4]),
+            open=Decimal(data[1]),
+            high=Decimal(data[2]),
+            low=Decimal(data[3]),
+            close=Decimal(data[4]),
             number_of_trades=int(data[8]),
-            base_asset_volume=float(data[5]),
-            quote_asset_volume=float(data[7]),
-            taker_buy_base_asset_volume=float(data[9]),
-            taker_buy_quote_asset_volume=float(data[10]),
+            base_asset_volume=Decimal(data[5]),
+            quote_asset_volume=Decimal(data[7]),
+            taker_buy_base_asset_volume=Decimal(data[9]),
+            taker_buy_quote_asset_volume=Decimal(data[10]),
             is_closed=True
         )
